+++ src/test/java/faang/school/postservice/service/comment/CommentServiceTest.java
@@ -1,9 +1,15 @@
 package faang.school.postservice.service.comment;
 
+import faang.school.postservice.client.CommentAnalyzer;
 import faang.school.postservice.client.UserServiceClient;
 import faang.school.postservice.dto.comment.CommentRequestDto;
 import faang.school.postservice.dto.comment.CommentResponseDto;
 import faang.school.postservice.dto.comment.CommentUpdateDto;
+import faang.school.postservice.dto.commentAnalyzer.response.AttributeScoreDto;
+import faang.school.postservice.dto.commentAnalyzer.response.SpanScoreDto;
+import faang.school.postservice.dto.commentAnalyzer.response.SummaryScoreDto;
+import faang.school.postservice.dto.commentAnalyzer.response.ToxicityScoreDto;
+import faang.school.postservice.enums.CommentToxicityType;
 import faang.school.postservice.mapper.CommentRequestMapper;
 import faang.school.postservice.mapper.CommentResponseMapper;
 import faang.school.postservice.model.Comment;
@@ -18,16 +24,22 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mapstruct.factory.Mappers;
 import org.mockito.ArgumentCaptor;
-import org.mockito.Captor;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.Spy;
 import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageImpl;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.test.util.ReflectionTestUtils;
+import reactor.core.publisher.Mono;
+import reactor.test.StepVerifier;
 
 import java.nio.charset.StandardCharsets;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 import static faang.school.postservice.contants.ErrorMessage.ERROR_NOT_AUTHOR_COMMENT;
@@ -42,6 +54,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.any;
 import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.times;
@@ -50,18 +63,15 @@ import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
 class CommentServiceTest {
-    private static final Long POST_ID = 1L;
-    private static final Long COMMENT_ID = 2L;
-    private static final Long AUTHOR_ID = 3L;
-
-    private static final String CONTENT = "Content";
-    private static final String UPDATE_CONTENT = "Update content";
 
     @InjectMocks
     private CommentServiceImpl commentService;
 
     @Mock
     private CommentRepository commentRepository;
+
+    @Mock
+    private CommentAnalyzer commentAnalyzer;
 
     @Mock
     private PostRepository postRepository;
@@ -75,14 +85,21 @@ class CommentServiceTest {
     @Spy
     private CommentRequestMapper commentRequestMapper = Mappers.getMapper(CommentRequestMapper.class);
 
-    @Captor
-    private ArgumentCaptor<Comment> commentCaptor;
+    private static final Long POST_ID = 1L;
+    private static final Long COMMENT_ID = 2L;
+    private static final Long AUTHOR_ID = 3L;
+
+    private static final String CONTENT = "Content";
+    private static final String UPDATE_CONTENT = "Update content";
 
     private CommentRequestDto commentRequestDto;
     private CommentResponseDto commentResponseDto;
     private CommentUpdateDto commentUpdateDto;
     private Comment comment;
     private Post post;
+    private Mono<ToxicityScoreDto> toxicityScore;
+    private final ArgumentCaptor<Comment> commentCaptor = ArgumentCaptor.forClass(Comment.class);
+    private final int commentModerationBatchSize = 3;
 
     @BeforeEach
     void setUp() {
@@ -107,6 +124,25 @@ class CommentServiceTest {
         commentUpdateDto = new CommentUpdateDto();
         commentUpdateDto.setAuthorId(AUTHOR_ID);
         commentUpdateDto.setContent(UPDATE_CONTENT);
+
+        ReflectionTestUtils.setField(commentService, "banBatchSize", 100);
+        ReflectionTestUtils.setField(commentService, "commentModerationTimeoutHours", 1);
+        ReflectionTestUtils.setField(commentService, "commentModerationBatchSize", commentModerationBatchSize);
+
+        toxicityScore = Mono.just(ToxicityScoreDto.builder()
+                .attributeScores(Map.of(
+                        CommentToxicityType.TOXICITY, new AttributeScoreDto(
+                                List.of(new SpanScoreDto(new SummaryScoreDto(0.15, "PROBABILITY"))),
+                                new SummaryScoreDto(0.15, "PROBABILITY")
+                        ),
+                        CommentToxicityType.INSULT, new AttributeScoreDto(
+                                List.of(new SpanScoreDto(new SummaryScoreDto(0.05, "PROBABILITY"))),
+                                new SummaryScoreDto(0.05, "PROBABILITY")
+                        )
+                ))
+                .languages(List.of("en"))
+                .detectedLanguages(List.of("en"))
+                .build());
     }
 
     //Positive
@@ -332,5 +368,53 @@ class CommentServiceTest {
 
         assertEquals(getErrorNotFoundComment(COMMENT_ID), exception.getMessage());
         verify(commentRepository, times(1)).findById(COMMENT_ID);
+    }
+
+    @Test
+    public void testModerateComments_moderationPassed() {
+        Comment comment1 = Comment.builder().content("content1").build();
+        Comment comment2 = Comment.builder().content("content2").build();
+        Comment comment3 = Comment.builder().content("content2").build();
+
+        List<Comment> comments = List.of(comment1, comment2, comment3);
+        Page<Comment> commentPage = new PageImpl<>(comments, PageRequest.of(0, commentModerationBatchSize), 3);
+
+        when(commentRepository.count()).thenReturn(3L);
+        when(commentRepository.findComments(any())).thenReturn(commentPage);
+        when(commentAnalyzer.analyzeComment(anyString())).thenReturn(toxicityScore);
+
+        Mono<Void> result = commentService.moderateComments();
+        StepVerifier.create(result)
+                .verifyComplete();
+
+        verify(commentRepository, times(3)).save(commentCaptor.capture());
+        List<Comment> capturedComments = commentCaptor.getAllValues();
+        assertEquals(3, capturedComments.size());
+        assertTrue(capturedComments.containsAll(comments));
+        assertTrue(capturedComments.stream().allMatch(Comment::isVerified));
+    }
+
+    @Test
+    public void testModerateComments_moderationFailed() {
+        Comment comment1 = Comment.builder().content("This text contains offensive content").build();
+        toxicityScore.block().getAttributeScores().get(CommentToxicityType.TOXICITY)
+                .setSummaryScore(new SummaryScoreDto(0.45, "PROBABILITY"));
+
+        List<Comment> comments = List.of(comment1);
+        Page<Comment> commentPage = new PageImpl<>(comments, PageRequest.of(0, 1), 1);
+
+        when(commentRepository.count()).thenReturn(1L);
+        when(commentRepository.findComments(any())).thenReturn(commentPage);
+        when(commentAnalyzer.analyzeComment(anyString())).thenReturn(toxicityScore);
+
+        Mono<Void> result = commentService.moderateComments();
+        StepVerifier.create(result)
+                .verifyComplete();
+
+        verify(commentRepository, times(1)).save(commentCaptor.capture());
+        List<Comment> capturedComments = commentCaptor.getAllValues();
+        assertEquals(1, capturedComments.size());
+        assertTrue(capturedComments.containsAll(comments));
+        assertFalse(capturedComments.stream().allMatch(Comment::isVerified));
     }
 }
