+++ src/test/java/faang/school/postservice/service/hashtags/HashtagRedisServiceTest.java
@@ -1,4 +1,4 @@
-package faang.school.postservice.hashtags;
+package faang.school.postservice.service.hashtags;
 
 import faang.school.postservice.model.Post;
 import faang.school.postservice.service.hashtags.HashtagRedisService;
