+++ src/test/java/faang/school/postservice/service/like/LikeServiceTest.java
@@ -1,4 +1,4 @@
-package faang.school.postservice.likeService;
+package faang.school.postservice.service.like;
 
 import faang.school.postservice.client.UserServiceClient;
 import faang.school.postservice.exception.LikeException;
